+++ packages/core/src/types/hook.ts
@@ -22,7 +22,7 @@ import {
 } from '@cosmjs/stargate';
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 
-import { ChainWalletBase } from '../bases';
+import { ChainWalletBase, MainWalletBase } from '../bases';
 import { NameService } from '../name-service';
 import { WalletRepo } from '../repository';
 import { ChainName, ChainRecord } from './chain';
@@ -137,6 +137,7 @@ export interface ChainContext extends ChainWalletContext {
 export interface ManagerContext {
   chainRecords: ChainRecord[];
   walletRepos: WalletRepo[];
+  mainWallets: MainWalletBase[];
   defaultNameService: NameServiceName;
   getChainRecord: (chainName: ChainName) => ChainRecord;
   getWalletRepo: (chainName: ChainName) => WalletRepo;
@@ -158,6 +159,7 @@ export interface ModalThemeContext {
 }
 
 export interface WalletContext {
+  mainWallet: MainWalletBase | undefined;
   chainWallets: ChainWalletBase[];
   wallet: Wallet | undefined;
   status: WalletStatus;