+++ packages/core/types/types/hook.d.ts
@@ -4,7 +4,7 @@ import { CosmWasmClient, SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate
 import { DirectSignResponse, EncodeObject, OfflineDirectSigner, OfflineSigner } from '@cosmjs/proto-signing';
 import { DeliverTxResponse, SigningStargateClient, StargateClient, StdFee } from '@cosmjs/stargate';
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
-import { ChainWalletBase } from '../bases';
+import { ChainWalletBase, MainWalletBase } from '../bases';
 import { NameService } from '../name-service';
 import { WalletRepo } from '../repository';
 import { ChainName, ChainRecord } from './chain';
@@ -61,6 +61,7 @@ export interface ChainContext extends ChainWalletContext {
 export interface ManagerContext {
     chainRecords: ChainRecord[];
     walletRepos: WalletRepo[];
+    mainWallets: MainWalletBase[];
     defaultNameService: NameServiceName;
     getChainRecord: (chainName: ChainName) => ChainRecord;
     getWalletRepo: (chainName: ChainName) => WalletRepo;
@@ -75,6 +76,7 @@ export interface ModalThemeContext {
     setModalTheme: (theme: ModalTheme) => void;
 }
 export interface WalletContext {
+    mainWallet: MainWalletBase | undefined;
     chainWallets: ChainWalletBase[];
     wallet: Wallet | undefined;
     status: WalletStatus;