+++ packages/core/types/types/wallet.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="react" />
 /// <reference types="node" />
 /// <reference types="long" />
 import { AccountData, AminoSignResponse, OfflineAminoSigner, StdSignature, StdSignDoc } from '@cosmjs/amino';