+++ packages/react/types/modal/components/views/QRCode.d.ts
@@ -1,2 +1,3 @@
+/// <reference types="react" />
 import { WalletViewProps } from '@cosmos-kit/core';
 export declare const QRCodeView: ({ onClose, onReturn, wallet }: WalletViewProps) => JSX.Element;