+++ packages/react/types/modal/components/views/WalletList.d.ts
@@ -1,2 +1,3 @@
+/// <reference types="react" />
 import { WalletListViewProps } from '@cosmos-kit/core';
 export declare const WalletListView: ({ onClose, wallets, initialFocus, }: WalletListViewProps) => JSX.Element;