+++ packages/react/types/modal/modal.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="react" />
 import { ModalViews, WalletModalProps } from '@cosmos-kit/core';
 export declare const DefaultModal: ({ isOpen, setOpen, walletRepo, }: WalletModalProps) => JSX.Element;
 export declare const WalletModal: ({ isOpen, setOpen, walletRepo, modalViews, includeAllWalletsOnMobile, }: WalletModalProps & {