+++ wallets/exodus-extension/src/config.ts
@@ -1,0 +1,7 @@
+import { EndpointOptions } from '@cosmos-kit/core';
+
+export const preferredEndpoints: EndpointOptions['endpoints'] = {
+  osmosis: {
+    rpc: ['https://atom4.a.exodus.io/api'],
+  },
+};