+++ wallets/exodus-extension/src/exodus.ts
@@ -1,0 +1,9 @@
+import { preferredEndpoints } from './config';
+import { exodusExtensionInfo, ExodusExtensionWallet } from './extension';
+
+const exodusExtension = new ExodusExtensionWallet(
+  exodusExtensionInfo,
+  preferredEndpoints
+);
+
+export const wallets = [exodusExtension];