+++ packages/example/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.45.15](https://github.com/cosmology-tech/cosmos-kit/compare/example@0.45.14...example@0.45.15) (2023-06-09)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [0.45.14](https://github.com/cosmology-tech/cosmos-kit/compare/example@0.45.13...example@0.45.14) (2023-06-09)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [0.45.13](https://github.com/cosmology-tech/cosmos-kit/compare/example@0.45.12...example@0.45.13) (2023-06-08)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [0.45.12](https://github.com/cosmology-tech/cosmos-kit/compare/example@0.45.11...example@0.45.12) (2023-06-08)
+
+**Note:** Version bump only for package example
+
+
+
+
+
 ## [0.45.11](https://github.com/cosmology-tech/cosmos-kit/compare/example@0.45.10...example@0.45.11) (2023-06-05)
 
 **Note:** Version bump only for package example