+++ wallets/compass-extension/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/compass-extension@0.1.2...@cosmos-kit/compass-extension@0.1.3) (2023-06-09)
+
+**Note:** Version bump only for package @cosmos-kit/compass-extension
+
+
+
+
+
+## [0.1.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/compass-extension@0.1.1...@cosmos-kit/compass-extension@0.1.2) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/compass-extension
+
+
+
+
+
+## [0.1.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/compass-extension@0.1.0...@cosmos-kit/compass-extension@0.1.1) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/compass-extension
+
+
+
+
+
 # 0.1.0 (2023-06-05)