+++ wallets/keplr-mobile/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.38](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-mobile@0.33.37...@cosmos-kit/keplr-mobile@0.33.38) (2023-06-09)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-mobile
+
+
+
+
+
+## [0.33.37](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-mobile@0.33.36...@cosmos-kit/keplr-mobile@0.33.37) (2023-06-09)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-mobile
+
+
+
+
+
+## [0.33.36](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-mobile@0.33.35...@cosmos-kit/keplr-mobile@0.33.36) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-mobile
+
+
+
+
+
+## [0.33.35](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-mobile@0.33.34...@cosmos-kit/keplr-mobile@0.33.35) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-mobile
+
+
+
+
+
 ## [0.33.34](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-mobile@0.33.33...@cosmos-kit/keplr-mobile@0.33.34) (2023-06-05)
 
 **Note:** Version bump only for package @cosmos-kit/keplr-mobile