+++ wallets/vectis-extension/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.10](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/vectis-extension@0.16.9...@cosmos-kit/vectis-extension@0.16.10) (2023-06-09)
+
+**Note:** Version bump only for package @cosmos-kit/vectis-extension
+
+
+
+
+
+## [0.16.9](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/vectis-extension@0.16.8...@cosmos-kit/vectis-extension@0.16.9) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/vectis-extension
+
+
+
+
+
+## [0.16.8](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/vectis-extension@0.16.7...@cosmos-kit/vectis-extension@0.16.8) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/vectis-extension
+
+
+
+
+
 ## [0.16.7](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/vectis-extension@0.16.6...@cosmos-kit/vectis-extension@0.16.7) (2023-06-05)
 
 **Note:** Version bump only for package @cosmos-kit/vectis-extension