+++ wallets/web3auth/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.17](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@0.0.16...@cosmos-kit/web3auth@0.0.17) (2023-06-09)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+## [0.0.16](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@0.0.15...@cosmos-kit/web3auth@0.0.16) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+## [0.0.15](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@0.0.14...@cosmos-kit/web3auth@0.0.15) (2023-06-08)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
 ## [0.0.14](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@0.0.13...@cosmos-kit/web3auth@0.0.14) (2023-06-05)
 
 **Note:** Version bump only for package @cosmos-kit/web3auth